[BE] 인가 기능 간소화 (#272)
* feat: 프로덕션 코드에서 인가 기능 삭제 #271

* test: 인가 기능 테스트 코드 삭제 #271

* test: 최소한의 인가 기능 구현 #271

* refactor: 사용하지 않는 memberId 파라미터 삭제 #271

+++ be/src/main/java/codesquad/bookkbookk/domain/bookmark/service/BookmarkService.java
@@ -12,10 +12,10 @@ import org.springframework.transaction.annotation.Transactional;
 import codesquad.bookkbookk.common.error.exception.BookmarkNotFoundException;
 import codesquad.bookkbookk.common.error.exception.BookmarkReactionExistsException;
 import codesquad.bookkbookk.common.error.exception.BookmarkReactionNotFoundException;
+import codesquad.bookkbookk.common.error.exception.MemberIsNotBookmarkWriterException;
 import codesquad.bookkbookk.common.error.exception.MemberNotFoundException;
 import codesquad.bookkbookk.common.error.exception.TopicNotFoundException;
 import codesquad.bookkbookk.common.type.Reaction;
-import codesquad.bookkbookk.domain.auth.service.AuthorizationService;
 import codesquad.bookkbookk.domain.bookmark.data.dto.BookmarkFilter;
 import codesquad.bookkbookk.domain.bookmark.data.dto.CreateBookmarkReactionRequest;
 import codesquad.bookkbookk.domain.bookmark.data.dto.CreateBookmarkRequest;
@@ -41,8 +41,6 @@ import lombok.RequiredArgsConstructor;
 @RequiredArgsConstructor
 public class BookmarkService {
 
-    private final AuthorizationService authorizationService;
-
     private final BookmarkRepository bookmarkRepository;
     private final MemberRepository memberRepository;
     private final TopicRepository topicRepository;
@@ -51,8 +49,6 @@ public class BookmarkService {
 
     @Transactional
     public void createBookmark(Long memberId, CreateBookmarkRequest createBookmarkRequest) {
-        authorizationService.authorizeBookClubMembershipByTopicId(createBookmarkRequest.getTopicId(), memberId);
-
         Member member = memberRepository.findById(memberId).orElseThrow(MemberNotFoundException::new);
         Topic topic = topicRepository.findById(createBookmarkRequest.getTopicId())
                 .orElseThrow(TopicNotFoundException::new);
@@ -71,27 +67,21 @@ public class BookmarkService {
     }
 
     @Transactional(readOnly = true)
-    public List<ReadBookmarkResponse> readBookmarks(Long memberId, Long topicId) {
-        authorizationService.authorizeBookClubMembershipByTopicId(topicId, memberId);
-
+    public List<ReadBookmarkResponse> readBookmarks(Long topicId) {
         List<Bookmark> bookmarks = bookmarkRepository.findAllByTopicId(topicId);
 
         return ReadBookmarkResponse.from(bookmarks);
     }
 
     @Transactional(readOnly = true)
-    public List<ReadBookmarkResponse> readBookmarksWithFilter(Long memberId, Long bookId, BookmarkFilter bookmarkFilter) {
-        authorizationService.authorizeBookClubMembershipByBookId(bookId, memberId);
-
+    public List<ReadBookmarkResponse> readBookmarksWithFilter(Long bookId, BookmarkFilter bookmarkFilter) {
         List<Bookmark> bookmarks = bookmarkRepository.findAllByFilter(bookId, bookmarkFilter);
 
         return ReadBookmarkResponse.from(bookmarks);
     }
 
     @Transactional(readOnly = true)
-    public ReadBookmarkSliceResponse readBookmarkSlices(Long memberId, Long topicId, Pageable pageable) {
-        authorizationService.authorizeBookClubMembershipByTopicId(topicId, memberId);
-
+    public ReadBookmarkSliceResponse readBookmarkSlices(Long topicId, Pageable pageable) {
         Slice<Bookmark> bookmarkSlice = bookmarkRepository.findSliceByTopicId(topicId, pageable);
         List<Long> bookmarkIds = bookmarkSlice.getContent().stream()
                 .map(Bookmark::getId)
@@ -104,24 +94,25 @@ public class BookmarkService {
 
     @Transactional
     public void updateBookmark(Long memberId, Long bookmarkId, UpdateBookmarkRequest updateBookmarkRequest) {
-        authorizationService.authorizeBookmarkWriter(memberId, bookmarkId);
-
         Bookmark bookmark = bookmarkRepository.findById(bookmarkId).orElseThrow(BookmarkNotFoundException::new);
+        if (!bookmarkRepository.existsByIdAndWriterId(bookmarkId, memberId)) {
+            throw new MemberIsNotBookmarkWriterException();
+        }
 
         bookmark.updateBookmark(updateBookmarkRequest);
     }
 
     @Transactional
     public void deleteBookmark(Long memberId, Long bookmarkId) {
-        authorizationService.authorizeBookmarkWriter(memberId, bookmarkId);
+        if (!bookmarkRepository.existsByIdAndWriterId(bookmarkId, memberId)) {
+            throw new MemberIsNotBookmarkWriterException();
+        }
 
         bookmarkRepository.deleteById(bookmarkId);
     }
 
     @Transactional
     public void createBookmarkReaction(Long memberId, Long bookmarkId, CreateBookmarkReactionRequest request) {
-        authorizationService.authorizeBookClubMembershipByBookmarkId(bookmarkId, memberId);
-
         Reaction reaction = Reaction.of(request.getReactionName());
         if (bookmarkReactionRepository.existsByBookmarkIdAndReactorIdAndReaction(bookmarkId, memberId, reaction)) {
             throw new BookmarkReactionExistsException();
@@ -146,9 +137,7 @@ public class BookmarkService {
     }
 
     @Transactional(readOnly = true)
-    public ReadReactionsResponse readBookmarkReactions(Long memberId, Long bookmarkId) {
-        authorizationService.authorizeBookClubMembershipByBookmarkId(bookmarkId, memberId);
-
+    public ReadReactionsResponse readBookmarkReactions(Long bookmarkId) {
         return ReadReactionsResponse.fromBookmarkReactions(bookmarkReactionRepository.findAllByBookmarkId(bookmarkId));
     }
